- Bump version to 1.4.0 - Implement utility module removing non class code from the library - Add functionality to write the output to a file
